extract HeaderViewHolder from Header model
